backstage/server/agent-api-maf/requirements.txt
agent-framework-core>=1.0.0rc5
agent-framework-anthropic>=1.0.0b260319
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.1
httpx>=0.28.0
